Music change size type

+++ src/com/cs490/boom/Music.java
@@ -3,32 +3,32 @@ package com.cs490.boom;
 public class Music {
 	private String name;
 	private String path;
-	private long size;
+	private int size;
 	private int preference;
 	//critical points
 	
-	public Music(String name, String path, long size) {
+	public Music(String name, String path, int size) {
 		this.name = name;
 		this.path = path;
 		this.size = size;
 		this.preference = 0;
 	}
 	
-        public void setName(String name) {
-            this.name = name;
-        }
+    public void setName(String name) {
+        this.name = name;
+    }
         
-        public void setPath(String path) {
-            this.path = path;
-        }
+    public void setPath(String path) {
+        this.path = path;
+    }
         
-        public void setSize(long size) {
-            this.size = size;
-        }
+    public void setSize(int size) {
+        this.size = size;
+    }
         
-        public void setPreference(int preference) {
-            this.preference = preference;
-        }
+    public void setPreference(int preference) {
+        this.preference = preference;
+    }
         
 	public String getName() {
 		return name;
@@ -38,7 +38,7 @@ public class Music {
 		return path;
 	}
 	
-	public long gerSize() {
+	public int gerSize() {
 		return size;
 	}
